requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
attrs==25.4.0
base58==2.1.1
bitarray==3.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.1.0
distro==1.9.0
docstring_parser==0.17.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.128.4
frozenlist==1.8.0
groq==1.0.0
grpcio==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-anthropic==1.3.2
langchain-core==1.2.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
multidict==6.7.1
neo4j==6.1.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parsimonious==0.10.0
propcache==0.4.1
protobuf==6.33.5
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
setuptools==82.0.0
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
starlette==0.52.1
tenacity==9.1.4
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
web3==7.14.1
websockets==11.0.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
